LPS-94301 mail-reader - classes in module that are not exported should be in internal package
