+++ modules/apps/mail-reader/mail-reader-web/src/main/java/com/liferay/mail/reader/web/internal/portlet/route/MailFriendlyURLMapper.java
@@ -12,19 +12,31 @@
  * details.
  */
 
-package com.liferay.mail.reader.web.upgrade.v1_0_0;
+package com.liferay.mail.reader.web.internal.portlet.route;
 
 import com.liferay.mail.reader.constants.MailPortletKeys;
-import com.liferay.portal.kernel.upgrade.BaseUpgradePortletId;
+import com.liferay.portal.kernel.portlet.DefaultFriendlyURLMapper;
+import com.liferay.portal.kernel.portlet.FriendlyURLMapper;
+
+import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Peter Fellwock
+ * @author Sergio González
  */
-public class UpgradePortletId extends BaseUpgradePortletId {
+@Component(
+	property = {
+		"com.liferay.portlet.friendly-url-routes=META-INF/friendly-url-routes/routes.xml",
+		"javax.portlet.name=" + MailPortletKeys.MAIL
+	},
+	service = FriendlyURLMapper.class
+)
+public class MailFriendlyURLMapper extends DefaultFriendlyURLMapper {
 
 	@Override
-	protected String[][] getRenamePortletIdsArray() {
-		return new String[][] {{"1_WAR_mailportlet", MailPortletKeys.MAIL}};
+	public String getMapping() {
+		return _MAPPING;
 	}
+
+	private static final String _MAPPING = "mail";
 
 }
