+++ modules/apps/mail-reader/mail-reader-web/src/main/resources/META-INF/resources/init.jsp
@@ -37,7 +37,7 @@ page import="com.liferay.mail.reader.model.MessagesDisplay" %><%@
 page import="com.liferay.mail.reader.service.AccountLocalServiceUtil" %><%@
 page import="com.liferay.mail.reader.service.FolderLocalServiceUtil" %><%@
 page import="com.liferay.mail.reader.service.MessageLocalServiceUtil" %><%@
-page import="com.liferay.mail.reader.web.util.MailManager" %><%@
+page import="com.liferay.mail.reader.web.internal.util.MailManager" %><%@
 page import="com.liferay.petra.string.StringBundler" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
 page import="com.liferay.portal.kernel.json.JSONArray" %><%@
